Moving prod_operation to arithmetics plugin
