+++ tests/unit/plugins/arithmetics/prod_operation.cpp
